// src/app/core/models/ms-audit-management/corrective-actions/correcctive-actions.ts
export interface MsAuditCorrectiveActionPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: MsauditCorrectiveAction[];
}

export interface MsauditCorrectiveAction{
    id: number;
    title: string;
    closed_by: CreatedBy
    description:string;
    reference_code:number;
    finding_id: number;
    start_date: string;
    target_date: string;
    responsible_user:ResponsibleUsers;
    documents: Documents[];
    created_by: CreatedBy;
    created_at: string;
    findings: CorrectiveActionFindings;
    ms_audit_findings:any;
    ms_audit_finding_id:number;
    ms_audit_finding_corrective_action_status: CorrectiveActionStatus
    ms_audit_finding_corrective_action_status_updates: CorrectiveActionStatusUpdates[]
    resolved_by :{
        created_at:string;
        comment: string;
        percentage: number;
        documents: Documents[];
        created_by :{
            id: number,
            first_name: string,
            last_name: string,
            email: string;
            mobile: number;
            status:{
                id:number;
            }
            designation: {
                title:string;
            },
            image_token: string;
            status_id : number;
            image:Image
        }
    }
}

    export interface CorrectiveActionStatusUpdates{
        comment: string;
        created_at: string;
        created_by: CreatedBy
        finding_corrective_action_id: number;
        ms_audit_finding_corrective_action_status_id: number;
        ms_audit_finding_corrective_action_update_documents: Documents[]
        id: number;
        percentage: number;
    }
    
    export interface CorrectiveActionFindings{
        id: number;
        title: string;
        audit_id: number;
        description: string;
        departments: Department[];
        divisions: Division[];
        sections: Section[]
        sub_sections: SubSection[]
        organizations: Subsidiary[]
        risk_rating: RiskRating;
        reference_code: string;
        documents: Documents[];
        findingStatus: FindingStatus;
        finding_category: FindingCategory;
    }
    
    export interface CorrectiveActionStatus{
        color_code: string;
        id: number;
        label: string;
        language: CorrectiveActionLanguage[]
        type: string;
    }
    
    export interface CorrectiveActionLanguage{
        code: string;
        id: number;
        is_primary: number;
        is_rtl: number;
        pivot: CorrectiveActionPivot
        status_id: number;
        title: string;
        type: string;
    }
    
    export interface CorrectiveActionPivot{
        finding_corrective_action_status_id: number, 
        language_id: number, 
        title: string;
    }
    export interface CreatedBy{
        designation: string;
        first_name: string;
        last_name: string;
        image:Image;
      }
    

    export interface ResponsibleUsers{
        id: number,
        first_name: string,
        last_name: string,
        email: string;
        mobile: number;
        status:{
            id:number;
        }
        designation: string,
        image_token: string;
        image:Image
      }
    
      export interface Image {
        title: String;
        url: string;
        thumbnail_url: string
        token: string;
        image_token:string;
        size: number;
        ext: string;
    }
    
    export interface Documents {
        id: number;
        token: string;
        title: string;
        ext: string;
        size: string;
        url: string;
        created_at:string;
        created_by:number;
        thumbnail_url:string;
        updated_at:string;
        updated_by:string;
    }